config.py
import os
from dotenv import load_dotenv

# Load environment variables
load_dotenv()

# API Configuration
GEMINI_API_KEY = os.getenv('GEMINI_API_KEY') or 'AIzaSyC7ofQbuPuV8RsTfCcL4Y6XjB6v23GUj0w'

# Chatbot Configuration
CHATBOT_NAME = "CNTT Advisor"
CHATBOT_DESCRIPTION = "Chuyên gia tư vấn cho sinh viên ngành Công nghệ Thông tin"

# System Prompt for the chatbot
SYSTEM_PROMPT = """
Bạn là một chuyên gia tư vấn học tập và nghề nghiệp cho sinh viên ngành Công nghệ Thông tin (CNTT). 
Nhiệm vụ của bạn là:

1. Tư vấn về các môn học trong chương trình CNTT
2. Hướng dẫn lộ trình học tập và phát triển kỹ năng
3. Tư vấn về cơ hội nghề nghiệp trong lĩnh vực CNTT
4. Giải đáp thắc mắc về công nghệ, lập trình, và các xu hướng IT
5. Đưa ra lời khuyên về việc chọn chuyên ngành và định hướng nghề nghiệp

Hãy trả lời một cách thân thiện, chuyên nghiệp và hữu ích. 
Sử dụng tiếng Việt trong giao tiếp và đưa ra các ví dụ cụ thể khi cần thiết.
"""

# Chat history file
CHAT_HISTORY_FILE = "chat_history.json"
